move functions into utils
