+++ adomvi/datasets/augmentation.py
@@ -5,98 +5,9 @@ import numpy as np
 from matplotlib import pyplot as plt
 from pathlib import Path
 
+from adomvi.utils import load_labels, convert_yolo_to_voc, convert_voc_to_yolo, save_yolo_labels
+
 LOG = logging.getLogger(__name__)
-
-
-def load_labels(label_path: Path):
-    """
-    Load labels from a YOLO format txt file.
-
-    Args:
-        label_path (Path): Path to the label file.
-
-    Returns:
-        list: List of bounding boxes.
-    """
-    with open(label_path, "r") as file:
-        labels = file.readlines()
-    bboxes = []
-    for label in labels:
-        parts = label.strip().split()
-        class_id = int(parts[0])
-        x_center, y_center, width, height = map(float, parts[1:])
-        bboxes.append([x_center, y_center, width, height, class_id])
-    return bboxes
-
-
-def yolo_to_voc(image_shape: np.ndarray, bboxes: list):
-    """
-    Convert labels from YOLO format to Pascal VOC format.
-
-    Args:
-        image_shape (tuple): Shape of the image in the format (height, width).
-        bboxes (list): List of bounding boxes in YOLO format.
-
-    Returns:
-        list: List of bounding boxes in Pascal VOC format.
-    """
-    h, w = image_shape[:2]
-    voc_bboxes = []
-    for bbox in bboxes:
-        x_center, y_center, width, height, class_id = bbox
-        x_min = (x_center - width / 2) * w
-        x_max = (x_center + width / 2) * w
-        y_min = (y_center - height / 2) * h
-        y_max = (y_center + height / 2) * h
-
-        x_min = max(0, x_min)
-        y_min = max(0, y_min)
-        x_max = min(w, x_max)
-        y_max = min(h, y_max)
-        voc_bboxes.append([x_min, y_min, x_max, y_max, class_id])
-
-    return voc_bboxes
-
-
-def voc_to_yolo(image_shape: np.ndarray, bboxes: list):
-    """
-    Convert labels from Pascal VOC format to YOLO format.
-
-    Args:
-        image_shape (tuple): Shape of the image in the format (height, width).
-        bboxes (list): List of bounding boxes in Pascal VOC format.
-
-    Returns:
-        list: List of bounding boxes in YOLO format.
-    """
-    h, w = image_shape[:2]
-    yolo_bboxes = []
-    for bbox in bboxes:
-        x_min, y_min, x_max, y_max, class_id = bbox
-        x_center = (x_min + x_max) / 2 / w
-        y_center = (y_min + y_max) / 2 / h
-        width = (x_max - x_min) / w
-        height = (y_max - y_min) / h
-        yolo_bboxes.append([class_id, x_center, y_center, width, height])
-    return yolo_bboxes
-
-
-def save_yolo_labels(label_path: Path, bboxes: list):
-    """
-    Save labels in YOLO format.
-
-    Args:
-        label_path (Path): Path to the label file.
-        bboxes (list): List of bounding boxes in YOLO format.
-    """
-    with open(label_path, "w") as file:
-        for bbox in bboxes:
-            class_id, x_center, y_center, width, height = bbox
-
-            # stop 6 digits after the decimal point
-            file.write(
-                f"{class_id} {x_center:.6f} {y_center:.6f} {width:.6f} {height:.6f}\n"
-            )
 
 
 def apply_augmentation(image_path: Path, label_path: Path, augmentation_func):
@@ -113,10 +24,10 @@ def apply_augmentation(image_path: Path, label_path: Path, augmentation_func):
     """
     image = cv2.imread(image_path)
     bboxes = load_labels(label_path)
-    voc_bboxes = yolo_to_voc(image.shape, bboxes)
+    voc_bboxes = convert_yolo_to_voc(image.shape, bboxes)
 
     aug_image, aug_bboxes = augmentation_func(image, voc_bboxes)
-    aug_bboxes = voc_to_yolo(image.shape, aug_bboxes)
+    aug_bboxes = convert_voc_to_yolo(image.shape, aug_bboxes)
 
     # Save augmented image
     ext = image_path.suffix
